Add "storage texture" polyfills to the webgpu backend
WebGPU requires support for a minimum of 8 storage buffers in any shader, but we still run on GL-backed WebGPU contexts, and some GL contexts don't support storage buffers in the vertex shader. Add the same "storage texture" fallback that we use in GL the webgpu impl.

Diffs=
9b7ba31d4 Add "storage texture" polyfills to the webgpu backend (#6497)
43dcd5636 Tooling tweaks for Chris's new laptop (#6496)

+++ renderer/d3d/pls_render_context_d3d_impl.cpp
@@ -516,7 +516,7 @@ protected:
     BufferRingD3D(PLSRenderContextD3DImpl* plsImpl,
                   size_t capacityInBytes,
                   UINT bindFlags,
-                  size_t elementSizeInBytes,
+                  UINT elementSizeInBytes,
                   UINT miscFlags) :
         BufferRing(capacityInBytes), m_gpuContext(plsImpl->gpuContext())
     {
@@ -572,7 +572,7 @@ class StructuredBufferRingD3D : public BufferRingD3D
 public:
     StructuredBufferRingD3D(PLSRenderContextD3DImpl* plsImpl,
                             size_t capacityInBytes,
-                            size_t elementSizeInBytes) :
+                            UINT elementSizeInBytes) :
         BufferRingD3D(plsImpl,
                       capacityInBytes,
                       D3D11_BIND_SHADER_RESOURCE,
